Keep interpolate_bad_pixels in the input dtype
Previously interpolate_bad_pixels allocated a float64 kernel and cast the
'good' mask via .astype(float), so scipy.ndimage.convolve promoted float32
image data to float64 in its intermediates. The final assignment to a
float32 data_interp downcast on store, hiding the cost from the on-disk
result but transiently allocating ~133 MB extra per 4080x4080 chip during
masters generation.

This was the only float64 promotion in the pipeline's active calculation
paths (verified by grep of np.float64 / dtype= / numpy default-promoting
calls across kpfpipe/modules/ and kpfpipe/utils/). Welford accumulators
in _compute_stats_from_stream stay float64 by design for stability over
many frames; their output is downcast at the stack_frames boundary.

Adds tests covering dtype preservation for float32 and float64 input,
plus correctness checks for the interpolated values and good-pixel
preservation.

+++ kpfpipe/utils/stats.py
@@ -95,14 +95,16 @@ def interpolate_bad_pixels(data, mask, method='local', fill_outside=True):
 
     data_interp = data.copy()
 
-    # local convolution-based method (assumes isolated bad pixels)
+    # local convolution-based method (assumes isolated bad pixels).
+    # Cast kernel and weight mask to the input dtype so scipy.convolve does
+    # not promote float32 image data to float64 in its intermediates.
     if method == 'local':
         kernel = np.array([[1,2,1],
                            [2,0,2],
-                           [1,2,1]], dtype=float) / 12.0
+                           [1,2,1]], dtype=data.dtype) / 12.0
 
         neighbor_sum = convolve(data * good, kernel, mode='mirror')
-        weight = convolve(good.astype(float), kernel, mode='mirror')
+        weight = convolve(good.astype(data.dtype), kernel, mode='mirror')
 
         valid = bad & (weight > 0)
         data_interp[valid] = neighbor_sum[valid] / weight[valid]
